// apps/api/src/common/enums/workflow-event-type.enum.ts
export enum WorkflowEventType {
  RUN_CREATED = 'RUN_CREATED',
  RUN_STARTED = 'RUN_STARTED',
  PROVIDER_SELECTED = 'PROVIDER_SELECTED',
  PROVIDER_REQUEST_SENT = 'PROVIDER_REQUEST_SENT',
  PROVIDER_RESPONSE_RECEIVED = 'PROVIDER_RESPONSE_RECEIVED',
  VALIDATION_STARTED = 'VALIDATION_STARTED',
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  RUN_COMPLETED = 'RUN_COMPLETED',
  RUN_FAILED = 'RUN_FAILED',
}